+++ src/routes/router.tsx
@@ -6,7 +6,7 @@ import RootLayout from '@/layouts/RootLayout';
 import HomePage from '@/pages/home/Home';
 import MyRolePage from '@/pages/my-role/My-Role';
 import MySpeakPage from '@/pages/my-speak/My-Speak';
-import MySpeakSettingPage from '@/pages/my-speak/Setting/Setting';
+import MySpeakSettingPage from '@/pages/my-speak/setting/Setting';
 
 const router = createBrowserRouter([
   {
